Port examples to v3

+++ examples/src/test/java/org/mapfish/print/ExamplesTest.java
@@ -46,7 +46,7 @@ import static org.junit.Assert.fail;
 public class ExamplesTest {
     public static final String DEFAULT_SPRING_XML = "classpath:mapfish-spring-application-context.xml";
 
-    private static final String REQUEST_DATA_FILE = "requestData.json";
+    private static final String REQUEST_DATA_FILE = "requestData(-.*)?.json";
     private static final String CONFIG_FILE = "config.yaml";
     @Autowired
     MapPrinter mapPrinter;
@@ -76,17 +76,24 @@ public class ExamplesTest {
         try {
             final File configFile = new File(example, CONFIG_FILE);
             this.mapPrinter.setConfiguration(configFile);
-            String requestData = Files.asCharSource(new File(example, REQUEST_DATA_FILE), Charset.forName(Constants.DEFAULT_ENCODING)).read();
-            final PJsonObject jsonSpec = MapPrinter.parseSpec(requestData);
-            //ByteArrayOutputStream out = new ByteArrayOutputStream();
             
-            File file = new File("/tmp/test-" + example.getName() + ".pdf");
-            FileOutputStream fs = new FileOutputStream(file);
-            
-            Map<String, String> headers = Maps.newHashMap();
-            this.mapPrinter.print(jsonSpec, fs, headers);
-            
-            fs.close();
+            for (File requestFile : Files.fileTreeTraverser().children(example)) {
+                if (requestFile.isFile() && requestFile.getName().matches(REQUEST_DATA_FILE)) {
+                    String requestData = Files.asCharSource(requestFile, Charset.forName(Constants.DEFAULT_ENCODING)).read();
+                    final PJsonObject jsonSpec = MapPrinter.parseSpec(requestData);
+                    //ByteArrayOutputStream out = new ByteArrayOutputStream();
+                    
+                    File file = new File("/tmp/test-" + example.getName()
+                            + requestFile.getName().replace("requestData", "").replace(".json", "")
+                            + ".pdf");
+                    FileOutputStream fs = new FileOutputStream(file);
+                    
+                    Map<String, String> headers = Maps.newHashMap();
+                    this.mapPrinter.print(jsonSpec, fs, headers);
+                    
+                    fs.close();
+                }
+            }
         } catch (Throwable e) {
             errors.put(example.getName(), e);
         }
